requirements.txt
agent-framework==1.0.0
pydantic>=2.0
pyyaml
tiktoken
rich
textual>=0.75.0
requests
certifi
markdownify
wcmatch
python-dotenv
opencv-python-headless
filetype
mcp[ws]>=1.26.0,<2
